Added basic gameplay to mini 1 house byOrlando scene
Since the model of the house is an obj  I did not had the time to make objects for each wall and stair so i just started adding a bunch of box coliiders to da house. Not the best way to do it but I think is midweek and we need to start moving fast.